pyproject.toml
[project]
name = "autoregressive-diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "torch>=2.7",
    "triton>=2.3.0",
    "awscli",
    "setuptools",
    "einops",
    "ipykernel",
    "ipywidgets",
    "matplotlib",
    "moviepy",
    "scipy",
    "opencv-python",
    "accelerate",
    "gymnasium[box2d,classic-control]",
    "diffusers>=0.32.1",
    "transformers",
    "datasets",
    "diffusers",
    "opencv-python",
    "h5py",
    "mosaicml-streaming",
    # "cosmos_tokenizer @ git+https://github.com/NVIDIA/Cosmos-Tokenizer",
    "lpips",
    "av"
]

[tool.setuptools.packages.find]
include = ["edm2", "edm2.*"]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"

[tool.uv]
prerelease = "allow"